+++ contact_search_form/i18n/en_US.po
@@ -1,0 +1,68 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* contact_search_form
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 11.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2018-07-02 12:15+0000\n"
+"PO-Revision-Date: 2018-07-02 12:15+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: contact_search_form
+#: model:ir.ui.menu,name:contact_search_form.menu_customer_info
+msgid "Contact Search"
+msgstr "Contact Search"
+
+#. module: contact_search_form
+#: model:ir.actions.act_window,name:contact_search_form.action_dpo_view_itis
+#: model:ir.ui.view,arch_db:contact_search_form.view_dpo_form_view_itis
+msgid "Customer Info"
+msgstr "Customer Info"
+
+#. module: contact_search_form
+#: model:res.groups,name:contact_search_form.group_dpo
+msgid "Data Protection Officer"
+msgstr "Data Protection Officer"
+
+#. module: contact_search_form
+#: model:ir.ui.view,arch_db:contact_search_form.view_dpo_form_view_itis
+msgid "SEARCH"
+msgstr "SEARCH"
+
+#. module: contact_search_form
+#: model:ir.ui.view,arch_db:contact_search_form.dpo_view_tree
+msgid "Search Contact"
+msgstr "Search Contact"
+
+#. module: contact_search_form
+#: model:ir.ui.view,arch_db:contact_search_form.search_line_tree_view_itis
+msgid "Search Results"
+msgstr "Search Results"
+
+#. module: contact_search_form
+#: model:ir.ui.view,arch_db:contact_search_form.view_dpo_form_view_itis
+msgid "Search Terms"
+msgstr "Search Terms"
+
+#. module: contact_search_form
+#: model:ir.ui.view,arch_db:contact_search_form.search_line_tree_view_itis
+msgid "VIEW RECORD"
+msgstr "VIEW RECORD"
+
+#. module: contact_search_form
+#: model:ir.model,name:contact_search_form.model_dpo_view
+msgid "dpo.view"
+msgstr "dpo.view"
+
+#. module: contact_search_form
+#: model:ir.model,name:contact_search_form.model_search_line
+msgid "search.line"
+msgstr "search.line"
+
