+++ contact_search_form/views/contact_search.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<odoo>
+
+    <record id="view_dpo_form_view_itis" model="ir.ui.view">
+        <field name="name">view.dpo.form.view.itis</field>
+        <field name="model">dpo.view</field>
+        <field name="arch" type="xml">
+            <form string="Customer Info">
+                <div class="oe_title">
+                    <h1>
+                        <field name="name" default_focus="1" placeholder="Search Terms" required="1"/>
+                    </h1>
+
+                </div>
+                <group>
+                    <field name="model_ids" widget="many2many_tags"/>
+                </group>
+                <br/>
+                <br/>
+                <center>
+                    <button name="search_string" type="object" string="SEARCH" class="btn-primary"/>
+                </center>
+                <br/>
+                <!--<group>-->
+                    <field name="search_lines"/>
+                <!--</group>-->
+            </form>
+
+        </field>
+    </record>
+
+    <record id="dpo_view_tree" model="ir.ui.view">
+        <field name="name">DPO Tree view</field>
+        <field name="model">dpo.view</field>
+        <field name="priority">1</field>
+        <field name="arch" type="xml">
+            <tree string="Search Contact">
+                <field name='name'/>
+                <field name="model_ids"/>
+                <field name="search_lines"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="action_dpo_view_itis" model="ir.actions.act_window">
+         <field name="name">Customer Info</field>
+         <field name="type">ir.actions.act_window</field>
+         <field name="res_model">dpo.view</field>
+         <field name="view_mode">tree,form</field>
+         <field name="view_id" ref="dpo_view_tree"/>
+         <!--<field name="target">new</field>-->
+         <field name="view_type">form</field>
+         <!--<field name="help" type="html">-->
+             <!--<p class="oe_view_nocontent_create">No Content yet!-->
+             <!--</p>-->
+         <!--</field>-->
+    </record>
+
+    <menuitem id="menu_customer_info"
+              action="action_dpo_view_itis"
+              name="Contact Search"
+              parent="contacts.res_partner_menu_config" sequence="3" groups="contact_search_form.group_dpo"/>
+
+    <record id="search_line_tree_view_itis" model="ir.ui.view">
+         <field name="name">search.line.tree.view.itis</field>
+         <field name="model">search.line</field>
+         <field name="arch" type="xml">
+             <tree string="Search Results" create="false" delete="false" edit="false">
+                 <field name="model_id"/>
+                 <field name="record_id"/>
+                 <field name="record_name"/>
+                 <field name="field_list"/>
+                 <button name="open_record" type="object" string="VIEW RECORD" class="btn-primary"/>
+             </tree>
+         </field>
+    </record>
+
+</odoo>
