+++ website_contact_extend/data/email_templates.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<odoo>
+
+    <record id="verification_email_template" model="mail.template">
+        <field name="name">Email verification link</field>
+        <field name="subject">mail verification link</field>
+        <field name="email_to">${(not object.partner_id and object.email_from)|safe}</field>
+        <field name="model_id" ref="crm.model_crm_lead"/>
+        <field name="auto_delete" eval="True"/>
+        <field name="body_html"><![CDATA[
+            <p>
+            % if object.partner_id
+            Dear <strong>${object.partner_id.name}</strong>,<br/>
+            % endif
+            % if not object.partner_id
+            Dear <strong>${object.email_from}</strong>,<br/>
+            % endif
+
+            Please click on below link to verify your email address <strong>${object.email_link}</strong>.<br/>
+
+            Thank you.
+            </p>
+        ]]>
+        </field>
+    </record>
+
+    <record id="confirmation_email_template" model="mail.template">
+        <field name="name">Email confirmation</field>
+        <field name="subject">Verified Email</field>
+        <field name="email_to">${user.email|safe}</field>
+        <field name="model_id" ref="base.model_res_partner"/>
+        <field name="auto_delete" eval="True"/>
+        <field name="body_html"><![CDATA[
+            <p>
+            Dear <strong>${user.partner_id.name}</strong>,<br/>
+
+            This email address <strong> ${object.email}</strong> has been verified by user.<br/>
+
+            Thank you.
+            </p>
+        ]]>
+        </field>
+    </record>
+
+</odoo>
