+++ website_contact_extend/views/res_partner.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<odoo>
+
+    <record id="view_partner_form_gdpr_inherit" model="ir.ui.view">
+        <field name="name">res.partner.form</field>
+        <field name="model">res.partner</field>
+        <field eval="1" name="priority"/>
+        <field name="inherit_id" ref="base.view_partner_form"/>
+        <field name="arch" type="xml">
+            <field name="website"  position="after">
+                <!--<field name="contact_type" />-->
+                <field name="phone_contact" readonly="True" />
+                <field name="email_contact" readonly="True" />
+                <field name="letter_contact" readonly="True" />
+                <field name="is_verified" readonly="True" />
+                <field name="last_updated" readonly="True" />
+            </field>
+        </field>
+    </record>
+
+</odoo>
